Move Form example to it's own component
